fixed but delete post with no image was not working

+++ app_package/main/routes.py
@@ -269,11 +269,14 @@ def rincon(rincon_name):
 
             # TODO: delete photo
             rincon_post = sess.get(RinconsPosts, formDict.get('btn_delete_post'))
-            post_image_path_and_name = os.path.join(current_app.static_folder, "rincon_files", f"{rincon.id}_{rincon.name_no_spaces}",rincon_post.image_file_name)
+
+
+            if rincon_post.image_file_name != None:
+                post_image_path_and_name = os.path.join(current_app.static_folder, "rincon_files", f"{rincon.id}_{rincon.name_no_spaces}",rincon_post.image_file_name)
             
-            print("post_image_path_and_name: ", post_image_path_and_name)
-            if os.path.exists(post_image_path_and_name):
-                os.remove(post_image_path_and_name)
+                print("post_image_path_and_name: ", post_image_path_and_name)
+                if os.path.exists(post_image_path_and_name):
+                    os.remove(post_image_path_and_name)
             sess.query(RinconsPosts).filter_by(id = formDict.get('btn_delete_post')).delete()
             sess.query(RinconsPostsLikes).filter_by(post_id = formDict.get('btn_delete_post')).delete()
             sess.query(RinconsPostsComments).filter_by(post_id = formDict.get('btn_delete_post')).delete()
